MAINT: Move misplaced genmod results files
