+++ statsmodels/genmod/tests/results/glmnet_r_results.py
