+++ statsmodels/genmod/tests/test_glm.py
@@ -1924,7 +1924,7 @@ class TestRegularized(object):
     def test_regularized(self):
 
         import os
-        from . import glmnet_r_results
+        from .results import glmnet_r_results
 
         for dtype in "binomial", "poisson":
